Move graph visualization to XY example
The graph is only really relevant and interesting for the XY animal
example. We can only graph things with 2 inputs anyway so this
isn't really relevant to include in the main neural network library.

This could apply to the XOR example but the graph isn't really interesting
and we already have that artifact in the readme.